requirements.txt
kafka-python
requests
pandas
pyspark==3.5.0
minio
streamlit
langchain
openai
qdrant-client
duckdb
pyiceberg
